Separate SSE and AVX
