+++ lib/TH/vector/SSE.c
@@ -4,7 +4,6 @@
 #include <intrin.h>
 #endif
 
-
 static void THDoubleVector_fill_SSE(double *x, const double c, const ptrdiff_t n) {
   ptrdiff_t i;
   ptrdiff_t off;
@@ -165,140 +164,3 @@ static void THFloatVector_mul_SSE(float *y, const float *x, const float c, const
   }
 }
 
-static void THDoubleVector_mul_AVX(double *y, const double *x, const double c, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256d YMM15 = _mm256_set_pd(c, c, c, c);
-  __m256d YMM0, YMM1, YMM2, YMM3, YMM4, YMM5;
-  for (i=0; i<=((n)-8); i+=8) {
-    YMM0 = _mm256_loadu_pd(x+i);
-    YMM1 = _mm256_loadu_pd(x+i+4);
-    YMM4 = _mm256_mul_pd(YMM0, YMM15);
-    YMM5 = _mm256_mul_pd(YMM1, YMM15);
-    _mm256_storeu_pd(y+i, YMM4);
-    _mm256_storeu_pd(y+i+4, YMM5);
-  }
-  for (; i<n; i++) {
-    y[i] = x[i] * c;
-  }
-}
-
-static void THFloatVector_mul_AVX(float *y, const float *x, const float c, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256 YMM15 = _mm256_set_ps(c, c, c, c, c, c, c, c);
-  __m256 YMM0, YMM1, YMM2, YMM3, YMM4, YMM5;
-  for (i=0; i<=((n)-16); i+=16) {
-    YMM0 = _mm256_loadu_ps(x+i);
-    YMM1 = _mm256_loadu_ps(x+i+8);
-    YMM4 = _mm256_mul_ps(YMM0, YMM15);
-    YMM5 = _mm256_mul_ps(YMM1, YMM15);
-    _mm256_storeu_ps(y+i, YMM4);
-    _mm256_storeu_ps(y+i+8, YMM5);
-  }
-  for (; i<n; i++) {
-    y[i] = x[i] * c;
-  }
-}
-
-static void THDoubleVector_add_AVX(double *y, const double *x, const double c, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256d YMM15 = _mm256_set_pd(c, c, c, c);
-  __m256d YMM0, YMM1, YMM2, YMM3, YMM4, YMM5;
-  for (i=0; i<=((n)-8); i+=8) {
-    YMM0 = _mm256_loadu_pd(x+i);
-    YMM1 = _mm256_loadu_pd(x+i+4);
-    YMM4 = _mm256_add_pd(YMM0, YMM15);
-    YMM5 = _mm256_add_pd(YMM1, YMM15);
-    _mm256_storeu_pd(y+i, YMM4);
-    _mm256_storeu_pd(y+i+4, YMM5);
-  }
-  for (; i<(n); i++) {
-    y[i] = x[i] + c;
-  }
-}
-
-static void THFloatVector_add_AVX(float *y, const float *x, const float c, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256 YMM15 = _mm256_set_ps(c, c, c, c, c, c, c, c);
-  __m256 YMM0, YMM1, YMM2, YMM3, YMM4, YMM5;
-  for (i=0; i<=((n)-16); i+=16) {
-    YMM0 = _mm256_loadu_ps(x+i);
-    YMM1 = _mm256_loadu_ps(x+i+8);
-    YMM4 = _mm256_add_ps(YMM0, YMM15);
-    YMM5 = _mm256_add_ps(YMM1, YMM15);
-    _mm256_storeu_ps(y+i, YMM4);
-    _mm256_storeu_ps(y+i+8, YMM5);
-  }
-  for (; i<(n); i++) {
-    y[i] = x[i] + c;
-  }
-}
-
-static void THDoubleVector_cdiv_AVX(double *z, const double *x, const double *y, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256d YMM0, YMM1, YMM2, YMM3;
-  for (i=0; i<=((n)-8); i+=8) {
-    YMM0 = _mm256_loadu_pd(x+i);
-    YMM1 = _mm256_loadu_pd(x+i+4);
-    YMM2 = _mm256_loadu_pd(y+i);
-    YMM3 = _mm256_loadu_pd(y+i+4);
-    YMM2 = _mm256_div_pd(YMM0, YMM2);
-    YMM3 = _mm256_div_pd(YMM1, YMM3);
-    _mm256_storeu_pd(z+i, YMM2);
-    _mm256_storeu_pd(z+i+4, YMM3);
-  }
-  for (; i<(n); i++) {
-    z[i] = x[i] / y[i];
-  }
-}
-
-static void THFloatVector_cdiv_AVX(float *z, const float *x, const float *y, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256 YMM0, YMM1, YMM2, YMM3;
-  for (i=0; i<=((n)-16); i+=16) {
-    YMM0 = _mm256_loadu_ps(x+i);
-    YMM1 = _mm256_loadu_ps(x+i+8);
-    YMM2 = _mm256_loadu_ps(y+i);
-    YMM3 = _mm256_loadu_ps(y+i+8);
-    YMM2 = _mm256_div_ps(YMM0, YMM2);
-    YMM3 = _mm256_div_ps(YMM1, YMM3);
-    _mm256_storeu_ps(z+i, YMM2);
-    _mm256_storeu_ps(z+i+8, YMM3);
-  }
-  for (; i<(n); i++) {
-    z[i] = x[i] / y[i];
-  }
-}
-
-static void THDoubleVector_div_AVX(double *y, const double *x, const double c, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256d YMM15 = _mm256_set_pd(c, c, c, c);
-  __m256d YMM0, YMM1, YMM2, YMM3, YMM4, YMM5;
-  for (i=0; i<=((n)-8); i+=8) {
-    YMM0 = _mm256_loadu_pd(x+i);
-    YMM1 = _mm256_loadu_pd(x+i+4);
-    YMM4 = _mm256_div_pd(YMM0, YMM15);
-    YMM5 = _mm256_div_pd(YMM1, YMM15);
-    _mm256_storeu_pd(y+i, YMM4);
-    _mm256_storeu_pd(y+i+4, YMM5);
-  }
-  for (; i<(n); i++) {
-    y[i] = x[i] / c;
-  }
-}
-
-static void THFloatVector_div_AVX(float *y, const float *x, const float c, const ptrdiff_t n) {
-  ptrdiff_t i;
-  __m256 YMM15 = _mm256_set_ps(c, c, c, c, c, c, c, c);
-  __m256 YMM0, YMM1, YMM2, YMM3, YMM4, YMM5;
-  for (i=0; i<=((n)-16); i+=16) {
-    YMM0 = _mm256_loadu_ps(x+i);
-    YMM1 = _mm256_loadu_ps(x+i+8);
-    YMM4 = _mm256_div_ps(YMM0, YMM15);
-    YMM5 = _mm256_div_ps(YMM1, YMM15);
-    _mm256_storeu_ps(y+i, YMM4);
-    _mm256_storeu_ps(y+i+8, YMM5);
-  }
-  for (; i<(n); i++) {
-    y[i] = x[i] / c;
-  }
-}
